Enhance wallet functionality with async initialization and improved data handling

+++ src/Angor/Avalonia/AngorApp/Sections/Wallet/Operate/BroadcastedTransactionDesign.cs
@@ -7,7 +7,7 @@ public class BroadcastedTransactionDesign : IBroadcastedTransaction
     public string Address { get; init; }
     public decimal FeeRate { get; set; }
     public decimal TotalFee { get; set; }
-    public uint Amount { get; init; }
+    public long Amount { get; init; }
     public string Path { get; init; }
     public int UtxoCount { get; init; }
     public string ViewRawJson { get; init; }
